need to test animatiosn... fixed some int/float stuff

+++ app/src/main/java/com/example/fundrawingsmegantherock/DrawView.java
@@ -123,10 +123,10 @@ public class DrawView extends View {
     }
 
 
-    public int getdY(){
+    public float getdY(){
         return dY;
     }
-    public void setdY(int dY){
+    public void setdY(float dY){
         this.dY = dY;
     }
 }
